Implement derive(SizeWith) using SizeWith recursively on fields. Fixes #12.
This change makes the implementation of derive(SizeWith) recursively call
the SizeWith implementation on all struct fields and return the sum. I
verified that the compiler winds up constant-folding it all down to a single
constant for the examples I tried. This should provide reliable sizes for
structs even if they're not repr(packed).